+++ changelog.md
@@ -1,16 +1,21 @@
 # Changelog
 
 ## 2020-04-26
-Official release!  Check out the [docs][docs] to get started.
+
+Official release! Check out the [docs][docs] to get started.
 
 ### `hast-util-annotate@0.1.1`
+
 ### `unified-doc-extract-text-offsets@0.1.1`
+
 ### `processor@0.1.1`
+
 ### `unified-doc-react@0.1.9`
+
 ### `text-parse@0.1.1`
 
 Note that any versions preceding these official versions probably have failing builds.
 
-
 <!-- Links -->
+
 [docs]: https://unified-doc.netlify.com/
