+++ docs/demos/comments.mdx
@@ -4,22 +4,26 @@ menu: Demos
 route: demos/comments
 ---
 
-import CommentsDemo from './src/comments';
+import CommentsDemo from "./src/comments";
 
 # Comments
 
 ## Description
+
 This demo shows how commenting (and other similar custom post-annotation rendering) can be implemented using just the `annotations` prop and rendering comment nodes relative to the annotated nodes.
 
 ## Implementation
-The comment data is mocked in a simple way using the array index as data.  This implementation uses an effect hook and vanilla DOM methods to render comment nodes.
 
-Since all annotated nodes contain the `data-id` attribute, you can select these nodes with the relevant document `querySelector`. Note that the `id` attribute is conveniently and *uniquely* provided for annotated nodes that have the `data-start="true"` attribute values.  This demo uses the `data-start`, `data-end` attributes to determine how comment nodes are placed.
+The comment data is mocked in a simple way using the array index as data. This implementation uses an effect hook and vanilla DOM methods to render comment nodes.
+
+Since all annotated nodes contain the `data-id` attribute, you can select these nodes with the relevant document `querySelector`. Note that the `id` attribute is conveniently and _uniquely_ provided for annotated nodes that have the `data-start="true"` attribute values. This demo uses the `data-start`, `data-end` attributes to determine how comment nodes are placed.
 
 ## Improvements
-You can improve this demo by formally building out a commenting workflow with relevant data.  Note that the `annotation` object could hold onto the `comments` data related to the annotation.  You can also improve the comment nodes with better CSS and design.
+
+You can improve this demo by formally building out a commenting workflow with relevant data. Note that the `annotation` object could hold onto the `comments` data related to the annotation. You can also improve the comment nodes with better CSS and design.
 
 ## Concepts
+
 - Custom post-annotation rendering with web technologies.
 - Annotation `data-` attributes.
 - Custom CSS.
