+++ docs/demos/ebook.mdx
@@ -4,24 +4,29 @@ menu: Demos
 route: demos/ebook
 ---
 
-import EbookDemo from './src/ebook';
+import EbookDemo from "./src/ebook";
 
 # Ebook
 
 ## Description
+
 An Ebook demo supporting simple bookmark management.
+
 - To create a bookmark, select some text in the document and confirm the action.
 - To remove a bookmark, click on the marked text in the document and confirm the action.
 - Configure the bookmark category to visually distinguish types of bookmarks.
 - Click on the bookmark tags to jump to the bookmark in the document.
 
 ## Implementation
-This demo is extremely simple and uses the `window.confirm` method to implement bookmark actions.  These actions are implemented using the `annotationCallbacks` and `onSelectText` props.
+
+This demo is extremely simple and uses the `window.confirm` method to implement bookmark actions. These actions are implemented using the `annotationCallbacks` and `onSelectText` props.
 
 ## Improvements
+
 You can create more interactive bookmark workflows by taking advantage of the mouse event returned by the relevant callback props (`annotationCallbacks`, `onSelectText`).
 
 ## Concepts
+
 - Creating annotations from text selections.
 - Customizable annotations.
 - Permalinks to annotated nodes.
