+++ docs/demos/src/comments.css
@@ -1,30 +1,30 @@
 .demo-comment {
-	--padding: 4px;
-	--size: 10px;
-	align-items: center;
-	border-radius: 50%;
-	background: #0b5fff;
-	color: white;
-	cursor: pointer;
-	display: flex;
-	font-size: 8px;
-	font-style: normal;
-	font-weight: bold;
-	height: var(--size);
-	justify-content: center;
-	padding: var(--padding);
-	position: absolute;
-	top: calc(-1 * (var(--size) + var(--padding)));
-	width: var(--size);
+  --padding: 4px;
+  --size: 10px;
+  align-items: center;
+  border-radius: 50%;
+  background: #0b5fff;
+  color: white;
+  cursor: pointer;
+  display: flex;
+  font-size: 8px;
+  font-style: normal;
+  font-weight: bold;
+  height: var(--size);
+  justify-content: center;
+  padding: var(--padding);
+  position: absolute;
+  top: calc(-1 * (var(--size) + var(--padding)));
+  width: var(--size);
 }
 .demo-comment:hover {
-	opacity: 0.8;
+  opacity: 0.8;
 }
 
 .demo-comment-start {
-	left: 0;
+  left: 0;
 }
 
 .demo-comment-end {
-	right: 0;
+  right: 0;
 }
