+++ docs/demos/src/diff.css
@@ -1,7 +1,7 @@
 .diff-added {
-	background: #acf2bd;
+  background: #acf2bd;
 }
 
 .diff-removed {
-	background: #fdb8c0;
+  background: #fdb8c0;
 }
