+++ docs/home/home.mdx
@@ -3,11 +3,12 @@ name: Home
 route: /
 ---
 
-import Home from './src/home';
+import Home from "./src/home";
 
 [unified][unified] document renderer for content.
 
 <Home />
 
 <!-- Links -->
+
 [unified]: https://unifiedjs.com/
