+++ docs/packages/packages.mdx
@@ -3,12 +3,15 @@ name: Packages
 route: packages
 ---
 
-import { Link } from 'gatsby';
+import { Link } from "gatsby";
 
 # Packages
 
 `unified-doc` packages:
+
 - <Link to="./packages/unified-doc">unified-doc</Link>
 - <Link to="./packages/unified-doc-parse-text">unified-doc-parse-text</Link>
 - <Link to="./packages/unified-doc-react">unified-doc-react</Link>
-- <Link to="./packages/unified-doc-util-annotate">unified-doc-util-annotate</Link>
+- <Link to="./packages/unified-doc-util-annotate">
+    unified-doc-util-annotate
+  </Link>
