+++ docs/packages/unified-doc-react/annotations.mdx
@@ -4,14 +4,14 @@ menu: unified-doc-react
 route: packages/unified-doc-react/annotations
 ---
 
-import { Link } from 'gatsby';
-import { useState } from 'react';
+import { Link } from "gatsby";
+import { useState } from "react";
 
-import AnnotationsExample from './src/annotations';
+import AnnotationsExample from "./src/annotations";
 
 # Annotations
 
-Annotating documents is easily supported by simply using the `annotations` prop.  You can control tooltip behaviors and display annotation labels.  Annotations support click and hover callbacks.
+Annotating documents is easily supported by simply using the `annotations` prop. You can control tooltip behaviors and display annotation labels. Annotations support click and hover callbacks.
 
 In the following example, you can also create new `annotations` by selecting text in the document. The [next section](./managing-annotations) provides an interactive example that covers this in more detail.
