+++ docs/packages/unified-doc-react/managing-annotations.mdx
@@ -4,12 +4,12 @@ menu: unified-doc-react
 route: packages/unified-doc-react/managing-annotations
 ---
 
-import ManagingAnnotationsExample from './src/managing-annotations';
+import ManagingAnnotationsExample from "./src/managing-annotations";
 
 # Managing Annotations
 
-It is common for applications to provide a way to manage annotations (create, remove, link).  The following example demonstrates how to use the `onSelectText` callback to create an annotation based on the text captured in the document.  You can also click on marked text in the document to confirm removing annotations.
+It is common for applications to provide a way to manage annotations (create, remove, link). The following example demonstrates how to use the `onSelectText` callback to create an annotation based on the text captured in the document. You can also click on marked text in the document to confirm removing annotations.
 
-This example also shows that annotations can be managed in a  *controlled* manner.  You can remove annotation tags to remove them in the documents.  Click on the annotation tag to jump/link to it in the document.
+This example also shows that annotations can be managed in a _controlled_ manner. You can remove annotation tags to remove them in the documents. Click on the annotation tag to jump/link to it in the document.
 
 <ManagingAnnotationsExample />
