+++ docs/packages/unified-doc-react/plugins.mdx
@@ -4,7 +4,7 @@ menu: unified-doc-react
 route: packages/unified-doc-react/plugins
 ---
 
-import PluginsExample from './src/plugins';
+import PluginsExample from "./src/plugins";
 
 # Plugins
