--- docs/packages/unified-doc-react/src/annotation-interface.tsx
@@ -1,7 +1,0 @@
-import { Annotation } from '@packages/unified-doc-react/src';
-
-function AnnotationInterface(_props: Annotation): JSX.Element {
-  return null;
-}
-
-export default AnnotationInterface;
