+++ docs/packages/unified-doc-react/styling.mdx
@@ -4,11 +4,11 @@ menu: unified-doc-react
 route: packages/unified-doc-react/styling
 ---
 
-import StylingExample from './src/styling';
+import StylingExample from "./src/styling";
 
 # Styling
 
-`unified-doc-react` renders a HTML document that maintains high fidelity with the source content.  The only modifications made to the document is semantic additions of `mark` nodes (with `data-start` and `data-end` attributes).  This simple approach allows easily customizing the document with familiar web technologies and CSS.
+`unified-doc-react` renders a HTML document that maintains high fidelity with the source content. The only modifications made to the document is semantic additions of `mark` nodes (with `data-start` and `data-end` attributes). This simple approach allows easily customizing the document with familiar web technologies and CSS.
 
 Apply custom styles with dropdown selector below, and view the custom document and annotation CSS in the right panel in the following example.
