+++ docs/quick-start.mdx
@@ -3,7 +3,7 @@ name: Quick Start
 route: /quick-start
 ---
 
-import { Link } from 'gatsby';
+import { Link } from "gatsby";
 
 # Quick Start
 
@@ -13,27 +13,38 @@ This progressive documentation will cover various programs (processors, renderer
 
 The documentation is organized in the following main sections:
 
-<Link to="/readme"><h3>Readme</h3></Link>
+<Link to="/readme">
+  <h3>Readme</h3>
+</Link>
 
-An overview of the [**unified-doc**][unified-doc] ecosystem, covering motivations for a unified document renderer.  Goals and architectural considerations are discussed towards building programs in the ecosystem.
+An overview of the [**unified-doc**][unified-doc] ecosystem, covering motivations for a unified document renderer. Goals and architectural considerations are discussed towards building programs in the ecosystem.
 
-<Link to="/demos/ebook"><h3>Demos</h3></Link>
+<Link to="/demos/ebook">
+  <h3>Demos</h3>
+</Link>
 
 A small collection of document-based applications showcasing diverse features that are easily implemented using the `unified-doc` spec.
 
-<Link to="/unified-doc-react/readme"><h3>unified-doc-react</h3></Link>
+<Link to="/unified-doc-react/readme">
+  <h3>unified-doc-react</h3>
+</Link>
 
 The official [React][react] implementation of the `unified-doc` spec.
 
-<Link to="/packages/unified-doc-util-annotate"><h3>Packages</h3></Link>
+<Link to="/packages/unified-doc-util-annotate">
+  <h3>Packages</h3>
+</Link>
 
 Various interoperable packages in the `unified-doc` ecosystem that support parsing, processing, and converting content into a unified rendered document.
 
-<Link to="/changelog"><h3>Changelog</h3></Link>
+<Link to="/changelog">
+  <h3>Changelog</h3>
+</Link>
 
 Stay up to date on any recent or breaking changes.
 
 <!-- Links -->
+
 [react]: https://github.com/facebook/react
 [unified]: https://unifiedjs.com/
 [unified-doc]: https://github.com/chrisrzhou/unified-doc
