+++ packages/unified-doc-parse-text/package.json
@@ -1,29 +1,29 @@
 {
-	"name": "unified-doc-parse-text",
-	"version": "1.0.0",
-	"description": "unified parser to parse text content into a hast tree with a single text node",
-	"main": "dist/index.js",
-	"module": "dist/index.module.js",
-	"source": "index.js",
-	"typings": "src/index.d.ts",
-	"keywords": [
-		"unified",
-		"doc",
-		"parse",
-		"text",
-		"document",
-		"content",
-		"unist",
-		"hast",
-		"html"
-	],
-	"author": "Chris Zhou <chrisrzhou@pm.me>",
-	"license": "MIT",
-	"repository": "https://github.com/chrisrzhou/unified-doc/tree/master/packages/unified-doc-parse-text",
-	"bugs": "https://github.com/chrisrzhou/unified-doc/issues",
-	"homepage": "https://unified-doc.netlify.com/",
-	"scripts": {
-		"build": "microbundle build",
-		"prepare": "yarn build"
-	}
+  "name": "unified-doc-parse-text",
+  "version": "1.0.0",
+  "description": "unified parser to parse text content into a hast tree with a single text node",
+  "main": "dist/index.js",
+  "module": "dist/index.module.js",
+  "source": "index.js",
+  "typings": "src/index.d.ts",
+  "keywords": [
+    "unified",
+    "doc",
+    "parse",
+    "text",
+    "document",
+    "content",
+    "unist",
+    "hast",
+    "html"
+  ],
+  "author": "Chris Zhou <chrisrzhou@pm.me>",
+  "license": "MIT",
+  "repository": "https://github.com/chrisrzhou/unified-doc/tree/master/packages/unified-doc-parse-text",
+  "bugs": "https://github.com/chrisrzhou/unified-doc/issues",
+  "homepage": "https://unified-doc.netlify.com/",
+  "scripts": {
+    "build": "microbundle build",
+    "prepare": "yarn build"
+  }
 }
