+++ packages/unified-doc-util-annotate/package.json
@@ -1,40 +1,40 @@
 {
-	"name": "unified-doc-util-annotate",
-	"version": "1.0.0",
-	"description": "hast utility to annotate text nodes",
-	"main": "dist/index.js",
-	"module": "dist/index.module.js",
-	"source": "index.js",
-	"typings": "src/index.d.ts",
-	"keywords": [
-		"unified",
-		"doc",
-		"util",
-		"annotate",
-		"document",
-		"content",
-		"unist",
-		"hast",
-		"html",
-		"text",
-		"annotation",
-		"highlight",
-		"redlining",
-		"utility"
-	],
-	"author": "Chris Zhou <chrisrzhou@pm.me>",
-	"license": "MIT",
-	"repository": "https://github.com/chrisrzhou/unified-doc/tree/master/packages/unified-doc-util-annotate",
-	"bugs": "https://github.com/chrisrzhou/unified-doc/issues",
-	"homepage": "https://unified-doc.netlify.com/",
-	"scripts": {
-		"build": "microbundle build",
-		"prepare": "yarn build"
-	},
-	"dependencies": {
-		"hastscript": "^5.1.2",
-		"tippy.js": "^6.1.0",
-		"unist-util-visit-parents": "^3.0.2",
-		"uuid": "^8.0.0"
-	}
+  "name": "unified-doc-util-annotate",
+  "version": "1.0.0",
+  "description": "hast utility to annotate text nodes",
+  "main": "dist/index.js",
+  "module": "dist/index.module.js",
+  "source": "index.js",
+  "typings": "src/index.d.ts",
+  "keywords": [
+    "unified",
+    "doc",
+    "util",
+    "annotate",
+    "document",
+    "content",
+    "unist",
+    "hast",
+    "html",
+    "text",
+    "annotation",
+    "highlight",
+    "redlining",
+    "utility"
+  ],
+  "author": "Chris Zhou <chrisrzhou@pm.me>",
+  "license": "MIT",
+  "repository": "https://github.com/chrisrzhou/unified-doc/tree/master/packages/unified-doc-util-annotate",
+  "bugs": "https://github.com/chrisrzhou/unified-doc/issues",
+  "homepage": "https://unified-doc.netlify.com/",
+  "scripts": {
+    "build": "microbundle build",
+    "prepare": "yarn build"
+  },
+  "dependencies": {
+    "hastscript": "^5.1.2",
+    "tippy.js": "^6.1.0",
+    "unist-util-visit-parents": "^3.0.2",
+    "uuid": "^8.0.0"
+  }
 }
