+++ packages/unified-doc-util-annotate/src/get-annotated-nodes.d.ts
@@ -1,6 +1,6 @@
 import { Node } from 'unist';
 
-import { Annotation, AnnotationCallbacks, Optional } from '.';
+import { Annotation, AnnotationCallbacks, Optional } from './types';
 
 interface AnnotationData {
   allAnnotations: { [key: string]: Annotation };
