+++ packages/unified-doc-util-annotate/src/validate-annotations.d.ts
@@ -1,4 +1,4 @@
-import { Annotation } from '.';
+import { Annotation } from './types';
 
 export default function validateAnnotations(
   annotations: Annotation[],
