+++ packages/unified-doc-util-annotate/src/validate-annotations.test.js
@@ -5,7 +5,7 @@ describe('validateAnnotations', () => {
     expect(validateAnnotations([])).toEqual([]);
   });
 
-  it('should not return same array for annotations that are already sorted', () => {
+  it('should return same array for annotations that are already sorted', () => {
     expect(
       validateAnnotations([{ id: 'a', startOffset: 0, endOffset: 40 }]),
     ).toEqual([{ id: 'a', startOffset: 0, endOffset: 40 }]);
@@ -28,17 +28,17 @@ describe('validateAnnotations', () => {
     expect(
       validateAnnotations([
         { id: 'a', startOffset: 60, endOffset: 80 },
-        { id: 'b', startOffset: 40, endOffset: 120 },
+        { id: 'b', startOffset: 40, endOffset: 80 },
         { id: 'c', startOffset: 120, endOffset: 140 },
         { id: 'd', startOffset: 0, endOffset: 40 },
         { id: 'e', startOffset: 20, endOffset: 50 },
-        { id: 'f', startOffset: 40, endOffset: 80 },
+        { id: 'f', startOffset: 40, endOffset: 120 },
       ]),
     ).toEqual([
       { id: 'd', startOffset: 0, endOffset: 40 },
       { id: 'e', startOffset: 20, endOffset: 50 },
-      { id: 'b', startOffset: 40, endOffset: 120 },
-      { id: 'f', startOffset: 40, endOffset: 80 },
+      { id: 'f', startOffset: 40, endOffset: 120 },
+      { id: 'b', startOffset: 40, endOffset: 80 },
       { id: 'a', startOffset: 60, endOffset: 80 },
       { id: 'c', startOffset: 120, endOffset: 140 },
     ]);
