+++ packages/unified-doc/src/index.css
@@ -1,17 +1,17 @@
-@import url('~tippy.js/dist/tippy.css');
+@import url("~tippy.js/dist/tippy.css");
 
 mark {
-	position: relative;
+  position: relative;
 }
 
 mark[data-start="true"]::before {
-	background-color: white;
-	color: gray;
-	content: attr(label);
-	font-size: 6px;
-	font-weight: normal;
-	position: absolute;
-	text-transform: uppercase;
-	top: -6px;
-	white-space: nowrap;
+  background-color: white;
+  color: gray;
+  content: attr(label);
+  font-size: 6px;
+  font-weight: normal;
+  position: absolute;
+  text-transform: uppercase;
+  top: -6px;
+  white-space: nowrap;
 }
