+++ src/readme.md
@@ -1,3 +1,1 @@
-This section provides source code for the monorepo and docs.
-
-Specifically, it is currently purely reserved to support [shadowing Gatsby themes](https://www.gatsbyjs.org/docs/themes/shadowing/).
+This section is set up to support [shadowing Gatsby themes](https://www.gatsbyjs.org/docs/themes/shadowing/) for the docs.
