+++ tsconfig.json
@@ -1,25 +1,19 @@
 {
-	"compilerOptions": {
-		"allowJs": true,
-		"allowSyntheticDefaultImports": true,
-		"baseUrl": ".",
-		"checkJs": true,
-		"esModuleInterop": true,
-		"jsx": "react",
-		"noEmit": true,
-		"paths": {
-			"@docs/*": ["./docs/*"],
-			"@packages/*": ["./packages/*"],
-			"@src/*": ["./src/*"]
-		},
-		"resolveJsonModule": true
-	},
-	"include": [
-		"docs",
-		"packages",
-	],
-	"exclude": [
-		"**/dist",
-		"**/node_modules"
-	]
+  "compilerOptions": {
+    "allowJs": true,
+    "allowSyntheticDefaultImports": true,
+    "baseUrl": ".",
+    "checkJs": true,
+    "esModuleInterop": true,
+    "jsx": "react",
+    "noEmit": true,
+    "paths": {
+      "@docs/*": ["./docs/*"],
+      "@packages/*": ["./packages/*"],
+      "@src/*": ["./src/*"]
+    },
+    "resolveJsonModule": true
+  },
+  "include": ["docs", "packages"],
+  "exclude": ["**/dist", "**/node_modules"]
 }
